api_test/backend/requirements.txt
fastapi
uvicorn[standard]
celery[redis]
redis
python-dotenv
deepagents>
langchain-google-genai
langgraph
tavily-python
reportlab
jinja2
arxiv
xmltodict
wikipedia
langchain-community
langdetect
google-genai
semanticscholar
python-pptx
python-docx
pillow
duckduckgo-search
boto3
botocore
sqlalchemy
psycopg2-binary